Native: dark mode, status editor, overnight-place pins, speed/theme fixes
Dark mode: new ThemeStore/ThemePreference, Theme.swift's tokens all
switched to adaptive(light:dark:) UIColor pairs matching styles.css's
[data-theme='dark'] block, PinpopApp wired to it instead of a forced
.preferredColorScheme(.light). Appearance now lives in Me as a
Day/Night/Auto three-tile picker instead of a map-screen toggle.

Me: status-emoji editor (AuthService.updateStatus, MeView.statusCard),
avatar's status badge moved to the top-right corner on this screen only.

Map: own-location pin no longer shows MapKit's system location dot
underneath it — camera switched from follow-mode (.userLocation) to a
manual one-time .region(...) centre, which was the actual source of
the stray purple/blue dot under the pin's tail.

Speed: CLLocation.speed is Doppler-derived but a low-confidence fix
(indoors, a shake reorienting the phone) could still report a "valid"
speed that was pure noise, e.g. a phantom 7 MPH while stationary. Now
gated on CoreLocation's own speedAccuracy, plus a second requirement
of two walking-pace-or-faster readings within 90 seconds before a
speed reaches the map or gets persisted for friends to see.

Theme refresh bug: Friends/Explore/Messages/Notifications/PersonCard
sheets kept showing the old palette until closed and reopened a
second time after a theme change. Forced a clean rebuild on every
open via .id(theme.preference); Me itself left alone since it already
updates live and a reset there would blow away in-progress edits.

New feature: overnight-place pins. Reuses significant_places, a table
that existed in the schema with zero references anywhere in src/ or
ios-native/ (dead scaffolding from an earlier pass). PlacesService
watches ordinary location fixes and, only during local 1am-5am,
records which grid cell (~111m) you're in as a night sample, once per
calendar night. 7 consecutive nights at the same cell promotes the
pin from "Night place" to "Home"; 15 days without a fresh sample and
it stops showing. Best-effort by nature -- no background job, so a
night the app never ran during that window just doesn't count. Grid
cell stored as Int (not Int64 -- PostgrestFilterValue isn't conformed
by Int64 in the pinned SDK version).

Also: current city name in the top bar (CLGeocoder, refreshed only
after a 2km move), and a new friends-can-read RLS policy on
significant_places (shares_location_with, same rule zones already
use) so a friend who already sees your live dot sees this too --
folded into backend/supabase/setup.sql plus its own migration file.

+++ ios-native/Sources/Services/PlacesService.swift
@@ -116,8 +116,8 @@ enum PlacesService {
             let kind: String
             let lat: Double
             let lng: Double
-            let cellLat: Int64
-            let cellLng: Int64
+            let cellLat: Int
+            let cellLng: Int
             let score: Double
             let lastSeenAt: Date
         }
@@ -153,6 +153,9 @@ enum PlacesService {
     /// ~111 m grid — coarse enough to absorb ordinary GPS drift around a
     /// building, fine enough to keep separate addresses apart. Mirrors the
     /// `cell_lat`/`cell_lng` bigint columns already reserved for this on
-    /// `significant_places`.
-    private static func cell(_ value: Double) -> Int64 { Int64((value * 1000).rounded()) }
+    /// `significant_places`. `Int`, not `Int64` — `PostgrestFilterValue`
+    /// (needed for the `.eq(...)` lookup below) isn't conformed by `Int64`
+    /// in the SDK version this project pins; `Int` is 64-bit on iOS anyway,
+    /// so nothing is actually lost.
+    private static func cell(_ value: Double) -> Int { Int((value * 1000).rounded()) }
 }
